requirements.txt
pyyaml==6.0
marshmallow-dataclass==8.5.8
pandas~=1.4.2
scikit-learn~=1.0.2
numpy~=1.22.3
omegaconf~=2.1.2
hydra-core==1.1.2
hydra_colorlog==1.1.0
pytest~=7.1.2